- kernel_sunos5.c, asn1.c: Fixes for 64-bit Solaris 7
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/branches/V3-line@2715 06827809-a52a-0410-b366-d66718629ded

+++ snmplib/asn1.c
@@ -585,7 +585,7 @@ asn_parse_length(u_char  *data,
 	*length = 0;  /* protect against short lengths */
 	memmove(length, data + 1, (int)lengthbyte);
 	*length = ntohl(*length);
-	*length >>= (8 * ((sizeof(int)) - lengthbyte));
+	*length >>= (8 * ((sizeof(u_long)) - lengthbyte));
 	return data + lengthbyte + 1;
     } else { /* short asnlength */
 	*length = (long)lengthbyte;
